requirements.txt
Django==1.11
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Markdown==2.6.11
numpy==1.14.2
pandas==0.22.0
psycopg2==2.7.4
PyJWT==1.6.0
python-dateutil==2.7.0
pytz==2018.3
six==1.11.0
